Egg_weight_lcd_tare
HX711 Load Cell Amplifier:
LOADCELL_DOUT_PIN (Data Output Pin): Connected to pin 16 on the ESP32. This pin receives the digital output signal from the load cell amplifier.
LOADCELL_SCK_PIN (Serial Clock Pin): Connected to pin 4 on the ESP32. This pin provides the clock signal for synchronous communication with the load cell amplifier.
Push Button:
BUTTON_PIN: Connected to pin 5 on the ESP32. This pin is used for the push button input. It's configured with an internal pull-up resistor, so when the button is pressed, it connects the pin to ground.
Liquid Crystal Display (LCD):
SDA, SCL: Connected to the respective SDA (data) and SCL (clock) pins of the ESP32 for I2C communication. The LCD's address is set to 0x27.